move coin selection policies under a name 'CoinSelection.Policy.*'
